+++ src/index.js
@@ -5,8 +5,29 @@ import { createStore } from 'redux';
 import ReactDOM from 'react-dom';
 import App from './components/App';
 import './assets/styles/styles.css';
-import reducer, { initialState } from './reducers/index';
+import reducer from './reducers/index';
 
+
+const newID = () => parseInt(Math.random() * 100, 10);
+const initialState = {
+  books: [
+    {
+      id: newID(),
+      title: 'The lord of the rings',
+      category: 'Sci-Fi',
+    },
+    {
+      id: newID(),
+      title: 'Nightmare on Stret Elmer',
+      category: 'Horror',
+    },
+    {
+      id: newID(),
+      title: 'Im Watching you',
+      category: 'Suspense',
+    },
+  ],
+};
 const store = createStore(reducer, initialState);
 
 ReactDOM.render(
