+++ src/reducers/books.js
@@ -4,6 +4,7 @@ const books = (state, action) => {
   const newState = { ...state };
   switch (action.type) {
     case 'CREATE_BOOK':
+      console.log('Soy create book acion');
       break;
     case 'DELETE_BOOK':
       break;
